изменены типы для цены и веса

+++ KursachOliaMarina/Models/Dish.cs
@@ -13,9 +13,9 @@ namespace KursachOliaMarina.Models
         // можно сделать категорию тоже цифрой, тогда легко будет отсортировать
         // 1 = первое, 2 = второе, 3 = салат, 4 = десерт, 5 = напиток
         // в таком порядке они и будут идти потом в меню, а не вперемешку
-        public float Price { get; set; }
-        public float Weight { get; set; }
-        public int Popularity { get; set;}
+        public decimal Price { get; set; }
+        public decimal Weight { get; set; }
+        public int Popularity { get; set; }
         public int Weekday { get; set; }
         public string Note { get; set; } // Рыбное\мясное\сладкое\молочное
